Ash.CiString (ash v3.4.8)

Represents a case insensitive string

While some data layers are aware of case insensitive string types, in order for values of this type to be used in other parts of Ash Framework, it has to be embedded in a module this allows us to implement the Comparable protocol for it.

For the type implementation, see Ash.Type.CiString

Types

t()

@type t() :: %Ash.CiString{
  case: nil | :lower | :upper,
  casted?: boolean(),
  string: String.t()
}

Functions

compare(left, right)

new(value, casing \\ nil)

sigil_i(value, mods)

Creates a case insensitive string

to_comparable_string(value)

Returns the downcased value, only downcasing if it hasn't already been done

value(ci_string)