Ash.SatSolver (ash v3.4.8)

Tools for working with the satsolver that drives filter subset checking (for authorization)

Functions

b(statement)
(macro)

Creates tuples of a boolean statement.

i.e b(1 and 2) #=> {:and, 1, 2}

contains?(l1, l2)

find_non_equal_overlap(expression)

left_excludes_right(left, right)

Returns b(not (left and right))

left_implies_right(left, right)

Returns b(not (left and not right))

lift_equals_out_of_in(expression)

mutually_exclusive(predicates, acc \\ [])

Returns a statement expressing that the predicates are mutually exclusive.

mutually_exclusive_and_collectively_exhaustive(predicates)

Returns a statement expressing that the predicates are mutually exclusive and collectively exhaustive.

mutually_inclusive(predicates, acc \\ [])

Returns a statement expressing that the predicates are mutually inclusive

right_excludes_left(left, right)

Returns b(not (right and left))

right_implies_left(left, right)

Returns b(not (right and not left))

solutions_to_predicate_values(solution, bindings)

solve_expression(cnf)

split_in_expressions(sub_expr, non_equal_overlap)

strict_filter_subset(filter, candidate)

Returns true if the candidate filter returns the same or less data than the filter

synonymous_relationship_paths?(left_resource, candidate, search, right_resource \\ nil)

Returns true if the relationship paths are synonymous from a data perspective

to_cnf(expression)

Transforms a statement to Conjunctive Normal Form(CNF), as lists of lists of integers.

transform(resource, expression)

Prepares a filter for comparison

transform_and_solve(resource, expression)

Calls transform/2 and solves the expression

unbind(expression, map)

Remaps integers back to clauses